Lambda  修改已知的问题 2023-02-28

+++ erp-model/erp-model-plm/src/main/java/com/erp/model/plm/dto/BasicCategoryDTO.java
@@ -39,7 +39,7 @@ public class BasicCategoryDTO {
     /**
      * 产品数量
      */
-    private Integer productQuantity;
+    private Long productQuantity;
 
     /**
      * 分类集合
